// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  ADMIN: 'ADMIN',
  SPEAKER: 'SPEAKER',
  MODERATOR: 'MODERATOR',
  COLLABORATOR: 'COLLABORATOR',
  PARTICIPANT: 'PARTICIPANT',
  PARTNER: 'PARTNER'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const UserStatus = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
  SUSPENDED: 'SUSPENDED',
  PENDING_VALIDATION: 'PENDING_VALIDATION'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const WebinarStatus = {
  SCHEDULED: 'SCHEDULED',
  ONGOING: 'ONGOING',
  COMPLETED: 'COMPLETED',
  CANCELED: 'CANCELED'
} as const

export type WebinarStatus = (typeof WebinarStatus)[keyof typeof WebinarStatus]


export const ApplicationType = {
  SPEAKER: 'SPEAKER',
  MODERATOR: 'MODERATOR',
  COLLABORATOR: 'COLLABORATOR',
  PARTNER: 'PARTNER',
  PARTICIPANT: 'PARTICIPANT'
} as const

export type ApplicationType = (typeof ApplicationType)[keyof typeof ApplicationType]


export const ApplicationStatus = {
  PENDING: 'PENDING',
  ACCEPTED: 'ACCEPTED',
  REJECTED: 'REJECTED'
} as const

export type ApplicationStatus = (typeof ApplicationStatus)[keyof typeof ApplicationStatus]


export const DocumentType = {
  LAW: 'LAW',
  DECREE: 'DECREE',
  UNIFORM_ACT: 'UNIFORM_ACT',
  CONVENTION: 'CONVENTION',
  COURT_DECISION: 'COURT_DECISION',
  OTHER: 'OTHER'
} as const

export type DocumentType = (typeof DocumentType)[keyof typeof DocumentType]


export const PersonalizedSupportType = {
  LEGAL: 'LEGAL',
  SCIENTIFIC: 'SCIENTIFIC',
  ENTREPRENEURIAL: 'ENTREPRENEURIAL',
  OTHER: 'OTHER'
} as const

export type PersonalizedSupportType = (typeof PersonalizedSupportType)[keyof typeof PersonalizedSupportType]


export const SupportType = {
  PRESENTATION: 'PRESENTATION',
  REFERENCE_DOCUMENT: 'REFERENCE_DOCUMENT',
  VIDEO: 'VIDEO',
  OTHER: 'OTHER'
} as const

export type SupportType = (typeof SupportType)[keyof typeof SupportType]


export const ModerationStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED'
} as const

export type ModerationStatus = (typeof ModerationStatus)[keyof typeof ModerationStatus]


export const RegistrationStatus = {
  CONFIRMED: 'CONFIRMED',
  CANCELED: 'CANCELED'
} as const

export type RegistrationStatus = (typeof RegistrationStatus)[keyof typeof RegistrationStatus]